+++ spec/chef/recipes/gitlab-pages_spec.rb
@@ -29,6 +29,13 @@ describe 'gitlab::gitlab-pages' do
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-admin-secret-path="/var/opt/gitlab/gitlab-pages/admin.secret"})
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-admin-unix-listener="/var/opt/gitlab/gitlab-pages/admin.socket"})
 
+      # By default pages access_control is disabled
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-client-id})
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-redirect-uri})
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-server})
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-client-secret})
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-secret})
+
       expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-listen-http})
       expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-listen-https})
       expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-root-cert})
@@ -43,6 +50,39 @@ describe 'gitlab::gitlab-pages' do
 
     it 'correctly renders the pages log run file' do
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/log/run").with_content(%r{exec svlogd -tt /var/log/gitlab/gitlab-pages})
+    end
+  end
+
+  context 'with access control without id and secret' do
+    before do
+      stub_gitlab_rb(
+        external_url: 'https://gitlab.example.com',
+        pages_external_url: 'https://pages.example.com',
+        gitlab_pages: {
+          log_verbose: true,
+          auth_secret: 'auth_secret',
+          access_control: true
+        }
+      )
+    end
+
+    it 'authorizes pages with gitlab' do
+      allow(GitlabPages).to receive(:authorize_with_gitlab) {
+        Gitlab['gitlab_pages']['gitlab_secret'] = 'app_secret'
+        Gitlab['gitlab_pages']['gitlab_id'] = 'app_id'
+      }
+
+      expect(chef_run).to run_ruby_block('authorize pages with gitlab')
+        .at_converge_time
+      expect(chef_run).to run_ruby_block('re-populate GitLab Pages configuration options')
+        .at_converge_time
+      expect(GitlabPages).to receive(:authorize_with_gitlab)
+
+      chef_run.ruby_block('authorize pages with gitlab').block.call
+      chef_run.ruby_block('re-populate GitLab Pages configuration options').block.call
+
+      expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-client-id=app_id})
+      expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-client-secret=app_secret})
     end
   end
 
@@ -66,9 +106,17 @@ describe 'gitlab::gitlab-pages' do
           admin_https_cert: '/etc/gitlab/pages-admin.crt',
           admin_https_key: '/etc/gitlab/pages-admin.key',
           admin_https_listener: 'localhost:2345',
-          log_verbose: true
+          log_verbose: true,
+          gitlab_id: 'app_id',
+          gitlab_secret: 'app_secret',
+          auth_secret: 'auth_secret',
+          access_control: true
         }
       )
+    end
+
+    it 'skip authorize pages with gitlab when id and secret exists' do
+      expect(chef_run).not_to run_ruby_block('authorize pages with gitlab')
     end
 
     it 'correctly renders the pages service run file' do
@@ -97,6 +145,11 @@ describe 'gitlab::gitlab-pages' do
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-admin-https-key="/etc/gitlab/pages-admin.key"})
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-admin-https-listener="localhost:2345"})
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-log-verbose})
+      expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-client-id=app_id})
+      expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-redirect-uri="https://projects.pages.example.com/auth"})
+      expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-server="https://gitlab.example.com"})
+      expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-client-secret=app_secret})
+      expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-secret=auth_secret})
     end
 
     it 'correctly renders the pages log run file' do
@@ -121,6 +174,29 @@ describe 'gitlab::gitlab-pages' do
       expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run")
       expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-artifacts-server=})
       expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-artifacts-server-timeout=})
+    end
+  end
+
+  context 'with access control disabled' do
+    before do
+      stub_gitlab_rb(
+        external_url: 'https://gitlab.example.com',
+        pages_external_url: 'https://pages.example.com',
+        gitlab_pages: { access_control: false }
+      )
+    end
+
+    it 'skip authorize pages with gitlab if access control disabled' do
+      expect(chef_run).not_to run_ruby_block('authorize pages with gitlab')
+    end
+
+    it 'correctly renders the pages service run file' do
+      expect(chef_run).to render_file("/opt/gitlab/sv/gitlab-pages/run")
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-client-id=})
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-redirect-uri=})
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-server=})
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-client-secret=})
+      expect(chef_run).not_to render_file("/opt/gitlab/sv/gitlab-pages/run").with_content(%r{-auth-secret=})
     end
   end
 end